Move nodeDB out of iavl_tree.go
